Moved tests to a new class
